requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/production.in
#
amqp==5.2.0
    # via kombu
arrow==1.3.0
    # via -r requirements/production.in
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via html-sanitizer
billiard==4.2.0
    # via celery
boto3==1.34.89
    # via django-storages
botocore==1.34.89
    # via
    #   boto3
    #   django-s3direct
    #   s3transfer
celery[redis]==5.4.0
    # via
    #   -r requirements/production.in
    #   django-celery-beat
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==42.0.5
    # via django-rest-knox
django==5.0.4
    # via
    #   -r requirements/production.in
    #   django-appconf
    #   django-celery-beat
    #   django-citext
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-localflavor
    #   django-redis
    #   django-rest-knox
    #   django-s3direct
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
    #   drf-standardized-errors
django-appconf==1.0.6
    # via django-imagekit
django-celery-beat==2.6.0
    # via -r requirements/production.in
django-citext==1.0.0
    # via -r requirements/production.in
django-extensions==3.2.3
    # via -r requirements/production.in
django-filter==24.2
    # via -r requirements/production.in
django-health-check==3.18.1
    # via -r requirements/production.in
django-imagekit==5.0.0
    # via -r requirements/production.in
django-localflavor==4.0
    # via -r requirements/production.in
django-redis==5.4.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
django-s3direct @ git+https://github.com/Saritasa/django-s3direct.git
    # via -r requirements/production.in
django-storages[boto3]==1.14.2
    # via -r requirements/production.in
django-timezone-field==6.1.0
    # via django-celery-beat
djangorestframework==3.15.1
    # via
    #   -r requirements/production.in
    #   django-rest-knox
    #   django-s3direct
    #   drf-spectacular
    #   drf-standardized-errors
drf-spectacular==0.27.2
    # via
    #   -r requirements/production.in
    #   drf-standardized-errors
drf-standardized-errors[openapi]==0.13.0
    # via -r requirements/production.in
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.0
    # via -r requirements/production.in
faker==24.11.0
    # via factory-boy
html-sanitizer==2.4.1
    # via -r requirements/production.in
idna==3.7
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-standardized-errors
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.7
    # via celery
lxml==5.2.1
    # via
    #   html-sanitizer
    #   lxml-html-clean
lxml-html-clean==0.1.1
    # via html-sanitizer
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via -r requirements/production.in
packaging==24.0
    # via
    #   pytest
    #   pytest-sugar
pilkit==3.0
    # via django-imagekit
pillow==10.3.0
    # via
    #   -r requirements/production.in
    #   pilkit
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.43
    # via click-repl
psycopg[binary]==3.1.18
    # via
    #   -r requirements/production.in
    #   django-citext
psycopg-binary==3.1.18
    # via psycopg
pycparser==2.22
    # via cffi
pygments==2.17.2
    # via rich
pyjwt==2.8.0
    # via -r requirements/production.in
pytest==7.4.4
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-lazy-fixtures
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.8.0
    # via -r requirements/production.in
pytest-lazy-fixtures==1.0.7
    # via -r requirements/production.in
pytest-sugar==1.0.0
    # via -r requirements/production.in
pytest-xdist==3.5.0
    # via -r requirements/production.in
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   botocore
    #   celery
    #   faker
    #   python-crontab
python-decouple==3.8
    # via -r requirements/production.in
python-stdnum==1.20
    # via django-localflavor
pyuwsgi==2.0.23.post0
    # via -r requirements/production.in
pyyaml==6.0.1
    # via drf-spectacular
redis==5.0.3
    # via
    #   celery
    #   django-redis
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via stripe
rich==13.7.1
    # via -r requirements/production.in
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.1
    # via boto3
sentry-sdk==1.45.0
    # via -r requirements/production.in
six==1.16.0
    # via python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.5.0
    # via django
stripe==9.3.0
    # via -r requirements/production.in
termcolor==2.4.0
    # via pytest-sugar
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.11.0
    # via
    #   psycopg
    #   stripe
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.1
    # via
    #   botocore
    #   requests
    #   sentry-sdk
uwsgitop==0.12
    # via -r requirements/production.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
